Use glibc's section flag injection scheme
The section is first declared with the section flags in a module level
assembly block. Then for each usage, we inject a comment character at
the end of the section name so that it reuses the given flags.

This solves most of section of alias '...' must match section of its
target problems with GCC 8.

We still face problems with alias pointing to alias and we do not manage
compilers quoting section names yet.