Factor tools to be more standalone
This is a major change to the contents of the tools/ directory. We now
have more fine-grained tools, with better names and more clearly defined
purposes. Static allocation metadata is undergoing a major overhaul, to
use a common 'metavector' spine whose entries correspond to existing
symbols -- symtab, dynsym -- or also 'extrasym' (additional symbols we
generate) and relocation information (left in the binary after linking
with -q). To any of these, the metavector entry may add type
information. As of this commit, most of these tools build and many work,
although metavector does not.