Partial refactor to pass through allocsite info to alloctypes
In the big tools/* refactoring, emission of the synthetic allocsite
types went missing. Since these are declared weak, we were seeing null
uniqtype pointers for those alloc sites. To fix this, we need to restore
the property that all typed allocsites have a DWARF-described type,
which was relaxed by 79f2b4b0 to ease problems with our roottypes.c
definitions lacking summary codes. This commit is a small refactoring to
make this need clear and prepare for reverting 79f2b4b0... it still
doesn't fix this problem.