Generate roottypes.c from in-memory DIEs
Instead of a manually authored roottypes.c, we now generate this from
synthetic DIEs, in a new tool (pervasive-types). This partially reverts
commit 79f2b4b08b7949481602efb03b2ab339858d99d4 (which removed code to
create these in-memory DIEs). Pervasive/root uniqtypes are now pretty
much first-class citizens (e.g. they have summary codes) and all
allocatable types can be represented with DIEs during tool processing.
This is so that we can fix the issue with allocsite-synthetic types,
which are not currently generated in this branch (but that fix is not
here yet).

Conflicts:
	tools/allocsites-info.cpp