uniqtypes.cpp: weed out like-named aliases of distinct but identical types
We can get 'the same' type in different CUs, but named differently e.g.
if it's an anonyous struct and the header file defining it is symlinked
from different places in the build tree. This can also happen
coincidentally. Different names means different types, but the summary
code may be the same (arguably it shouldn't be... hmm).

Sometimes these distinct types get aliased to the same name (e.g. a
typedef), and because their codes are also identical, we end up trying
to emit multiple codeful aliases with distinct targets.

This change prunes these and prints a warning. It is currently broken in
a specific case (that was already a problem): multiple aliases that
name-mangle to the same string, e.g. "long int" (in DWARF) and
"long_int" (in C code). These cause an assertion failure, so we'll need
to make the name-mangling function actually injective.